Write recipe closer to pangeo-forge style
We were trying to figure out how the pangeo-forge-runner injects runner config
into classes like `StoreToZarr`. We thought there was an injection occuring that
relied on the pattern of defining a dataclass that inherits from PTransform.

We ended up finding that the `pangeo-forge-runner` code is very specifically
written to inject this config into the Zarr writers, and nothing else.